feat: clean structure and folders
