--- client/src/App.css
@@ -1,7 +1,0 @@
-.App {
-	text-align: center;
-}
-
-.calendar span{
-	color:red
-}
