+++ client/src/assets/API/api.js
@@ -1,0 +1,4 @@
+
+const API = "http://localhost:3000/api";
+
+export default API;
