+++ client/src/components/Footer/Footer.jsx
@@ -1,0 +1,1 @@
+// footer
