+++ client/src/components/Headers/EventHeader/EventHeader.jsx
@@ -1,0 +1,1 @@
+// Header del evento
