+++ client/src/components/ListEvents/ListEvents.jsx
@@ -1,0 +1,1 @@
+// lista de eventos
