+++ client/src/components/ListEvents/ListEvents.module.css
@@ -1,0 +1,1 @@
+/* estilo */
