+++ client/src/components/Navbar/Navbar.module.css
@@ -1,0 +1,1 @@
+/* estilo */
