+++ client/src/components/User/SignUp.jsx
@@ -1,0 +1,1 @@
+// registro
