+++ client/src/pages/EventDetails/EventDetails.jsx
@@ -1,0 +1,1 @@
+// detalles del evento con su header personalizado
