+++ client/src/pages/EventDetails/EventDetails.module.css
@@ -1,0 +1,1 @@
+/* estilo de la page  */
