+++ client/src/pages/Home/Home.jsx
@@ -1,0 +1,7 @@
+// home con sus componentes
+const Home = () => {
+  return (
+    <div>Home</div>
+  )
+}
+export default Home
