+++ client/src/pages/Home/Home.module.css
@@ -1,0 +1,1 @@
+/* estilo de home */
